refactor: CustomException을 Enumd으로 관리해서 중복 코드 줄이도록 변경
