--- backend/pcloud-api/src/main/java/com/api/config/interceptor/auth/LoginValidCheckerInterceptor.java
@@ -1,32 +1,0 @@
-package com.api.config.interceptor.auth;
-
-import com.api.config.interceptor.auth.support.AuthenticationContext;
-import com.api.config.interceptor.auth.support.AuthenticationExtractor;
-import com.common.auth.TokenProvider;
-import com.common.exception.LoginInvalidException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.HandlerInterceptor;
-
-@RequiredArgsConstructor
-@Component
-public class LoginValidCheckerInterceptor implements HandlerInterceptor {
-
-    private final TokenProvider tokenProvider;
-    private final AuthenticationContext authenticationContext;
-
-    @Override
-    public boolean preHandle(final HttpServletRequest request,
-                             final HttpServletResponse response,
-                             final Object handler) throws Exception {
-        String token = AuthenticationExtractor.extract(request)
-                .orElseThrow(LoginInvalidException::new);
-
-        Long memberId = tokenProvider.extract(token);
-        authenticationContext.setAuthentication(memberId);
-
-        return true;
-    }
-}
