--- backend/pcloud-api/src/main/java/com/api/config/interceptor/auth/support/HttpMethod.java
@@ -1,20 +1,0 @@
-package com.api.config.interceptor.auth.support;
-
-public enum HttpMethod {
-
-    GET,
-    POST,
-    PUT,
-    PATCH,
-    DELETE,
-    OPTIONS,
-    HEAD,
-    TRACE,
-    CONNECT,
-    ANY;
-
-    public boolean matches(final String pathMethod) {
-        return this == ANY ||
-                this.name().equalsIgnoreCase(pathMethod);
-    }
-}
