--- backend/pcloud-api/src/main/java/com/api/config/moduleutils/ComponentScanConfig.java
@@ -1,9 +1,0 @@
-package com.api.config.moduleutils;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(value = {"com.common", "com.domain", "com.infrastructure"})
-public class ComponentScanConfig {
-}
