+++ backend/pcloud-api/src/main/java/com/api/global/config/interceptor/auth/ParseMemberIdFromTokenInterceptor.java
@@ -1,0 +1,29 @@
+package com.api.global.config.interceptor.auth;
+
+import com.api.global.config.interceptor.auth.support.AuthenticationContext;
+import com.api.global.config.interceptor.auth.support.AuthenticationExtractor;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+@RequiredArgsConstructor
+@Component
+public class ParseMemberIdFromTokenInterceptor implements HandlerInterceptor {
+
+    private final LoginValidCheckerInterceptor loginValidCheckerInterceptor;
+    private final AuthenticationContext authenticationContext;
+
+    @Override
+    public boolean preHandle(final HttpServletRequest request,
+                             final HttpServletResponse response,
+                             final Object handler) throws Exception {
+        if (AuthenticationExtractor.extract(request).isEmpty()) {
+            authenticationContext.setAnonymous();
+            return true;
+        }
+
+        return loginValidCheckerInterceptor.preHandle(request, response, handler);
+    }
+}
