+++ backend/pcloud-api/src/main/java/com/api/global/config/moduleutils/PropertySourceScanConfig.java
@@ -1,0 +1,16 @@
+package com.api.global.config.moduleutils;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@PropertySource(
+        value = {
+                "classpath:application-common-${spring.profiles.active}.yml",
+                "classpath:application-domain-${spring.profiles.active}.yml",
+                "classpath:application-infrastructure-${spring.profiles.active}.yml"
+        },
+        factory = YamlPropertySourceFactory.class
+)
+@Configuration
+public class PropertySourceScanConfig {
+}
