+++ backend/pcloud-api/src/main/java/com/api/global/exception/response/ExceptionResponse.java
@@ -1,0 +1,7 @@
+package com.api.global.exception.response;
+
+public record ExceptionResponse(
+        String code,
+        String message
+) {
+}
