+++ backend/pcloud-api/src/test/java/com/api/config/interceptor/auth/support/PathRequestTest.java
@@ -1,5 +1,7 @@
 package com.api.config.interceptor.auth.support;
 
+import com.api.global.config.interceptor.auth.support.HttpMethod;
+import com.api.global.config.interceptor.auth.support.PathRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
