+++ backend/pcloud-common/src/main/java/com/common/exception/CustomException.java
@@ -1,0 +1,10 @@
+package com.common.exception;
+
+public abstract class CustomException extends RuntimeException {
+
+    protected CustomException(final CustomExceptionType customExceptionType) {
+        super("[%s]: %s".formatted(customExceptionType.name(), customExceptionType.getMessage()));
+    }
+
+    public abstract CustomExceptionType getExceptionType();
+}
