+++ backend/pcloud-common/src/main/java/com/common/exception/CustomExceptionType.java
@@ -1,0 +1,10 @@
+package com.common.exception;
+
+public interface CustomExceptionType {
+
+    String getMessage();
+
+    String name();
+
+    int getHttpStatusCode();
+}
