--- backend/pcloud-common/src/main/java/com/common/exception/ExpiredTokenException.java
@@ -1,8 +1,0 @@
-package com.common.exception;
-
-public class ExpiredTokenException extends RuntimeException {
-
-    public ExpiredTokenException() {
-        super("이미 만료된 토큰입니다");
-    }
-}
