--- backend/pcloud-common/src/main/java/com/common/exception/SignatureInvalidException.java
@@ -1,8 +1,0 @@
-package com.common.exception;
-
-public class SignatureInvalidException extends RuntimeException {
-
-    public SignatureInvalidException() {
-        super("서명을 확인하지 못했습니다.");
-    }
-}
