--- backend/pcloud-common/src/main/java/com/common/exception/TokenFormInvalidException.java
@@ -1,8 +1,0 @@
-package com.common.exception;
-
-public class TokenFormInvalidException extends RuntimeException {
-
-    public TokenFormInvalidException() {
-        super("토큰의 길이 및 형식이 올바르지 않습니다.");
-    }
-}
