--- backend/pcloud-common/src/main/java/com/common/exception/TokenInvalidException.java
@@ -1,8 +1,0 @@
-package com.common.exception;
-
-public class TokenInvalidException extends RuntimeException {
-
-    public TokenInvalidException() {
-        super("토큰이 유효하지 않습니다.");
-    }
-}
