--- backend/pcloud-common/src/main/java/com/common/exception/UnsupportedTokenException.java
@@ -1,8 +1,0 @@
-package com.common.exception;
-
-public class UnsupportedTokenException extends RuntimeException {
-
-    public UnsupportedTokenException() {
-        super("지원하지 않는 토큰입니다.");
-    }
-}
