--- backend/pcloud-domain/src/main/java/com/domain/domains/member/exception/MemberAlreadyExistedException.java
@@ -1,8 +1,0 @@
-package com.domain.domains.member.exception;
-
-public class MemberAlreadyExistedException extends RuntimeException {
-
-    public MemberAlreadyExistedException() {
-        super("Member가 이미 존재합니다.");
-    }
-}
