--- backend/pcloud-domain/src/main/java/com/domain/domains/member/exception/MemberNotFoundException.java
@@ -1,8 +1,0 @@
-package com.domain.domains.member.exception;
-
-public class MemberNotFoundException extends RuntimeException {
-
-    public MemberNotFoundException() {
-        super("Member를 찾을 수 없습니다.");
-    }
-}
