--- backend/pcloud-domain/src/main/java/com/domain/domains/member/exception/PasswordInvalidException.java
@@ -1,8 +1,0 @@
-package com.domain.domains.member.exception;
-
-public class PasswordInvalidException extends RuntimeException {
-
-    public PasswordInvalidException() {
-        super("패스워드가 일치하지 않습니다.");
-    }
-}
